requirements.txt
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
cachetools==5.3.0
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
comtypes==1.1.11
cryptography==37.0.2
Cython==0.29.28
decorator==4.4.2
Django==4.0.4
et-xmlfile==1.1.0
Flask==2.1.2
Flask-Cors==3.0.10
google-api-core==2.11.0
google-api-python-client==2.75.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.58.0
googletrans==3.1.0a0
gspread==5.7.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.7.10
httpcore==0.9.1
httplib2==0.21.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
image==1.5.33
imageio==2.19.3
imageio-ffmpeg==0.4.7
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
MouseInfo==0.1.3
moviepy==1.0.3
mysql-connector-python==8.0.29
numpy==1.23.0
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.0
outcome==1.2.0
packaging==21.3
pandas==1.5.3
patool==1.12
Pillow==9.1.0
proglog==0.1.10
progressbar==2.5
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///C:/Users/shahin-zt381/Downloads/PyAudio-0.2.11-cp39-cp39-win_amd64.whl
PyAutoGUI==0.9.53
pybrowsers==0.4.1
pycparser==2.21
pydub==0.25.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyOpenSSL==22.0.0
pyparsing==3.0.8
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
pytesseract @ git+https://github.com/madmaze/pytesseract.git@a4c4f5faca5aaa18b3bcf5a5c86282948f6fdc45
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
pyttsx3==2.90
pytweening==1.0.4
pytz==2022.7.1
pywhatkit==5.4
pywin32==303
rarfile==4.0
requests==2.28.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
sqlparse==0.4.2
tesseract==0.1.3
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
webdriver-manager==3.8.0
Werkzeug==2.1.2
wikipedia==1.4.0
wsproto==1.1.0
zipp==3.8.0
